Tuffy experiment infrastructure weight writing.
There was an issue where if the learned weight was written in exponential notation then the awk script in the tuffy helper file would not increment the positional counter correctly and write the weight in the wrong location. Also, the output of tuffy's built in learner is not written in syntax that tuffy can then use for inference so the the weights need to be written over to a new model file. But the weights are also not printed so they align with the model weights in general. They provide a key but the key can have a decimal for any rules with a constant. This decimal was not being used before and this causes the weights to be shared across rules when their was indeed a unique weight learned by the weight learner. The solution is adhoc and a general solution is difficult to come up with.